Removes PortMidi
This turns out to be unavoidable. To get proper device detection we
have to call MIDIClientCreate which turns out to be called by PortMidi
at initialization time. I will reimplement the rest of the Fa.Midi
module in this branch and then think about abstracting it out to
the platform directory.