Scheduling drifts
This bug was hard to track down. Appeared on large time values (from clocks) or when complex
fractions was being used in the scheduler. The cause turned out to be an overflow in
fa_ratio_add. We might want to think more about how to do arithmetic on time etc.

Fixed by adding a safe version of ratio_add that falls back to floating-point for
complex fractions. This is sufficient for the scheduler (which does not do anything
but adding), but possibly not for other uses.

Fixes #127